+++ tests/inspectors/main.rs
@@ -1,1 +1,1 @@
-mod external_functions;
+mod function_source;